Move PromptTokenizer into the prompt-tokenizer module
