+++ prompt/prompt-tokenizer/build.gradle.kts
@@ -13,6 +13,7 @@ kotlin {
         commonMain {
             dependencies {
                 api(project(":prompt:prompt-llm"))
+                api(project(":prompt:prompt-model"))
                 api(libs.kotlinx.serialization.json)
                 api(libs.kotlinx.datetime)
             }
